+++ src/main/java/com/andersmmg/create_alerted/Config.java
@@ -11,7 +11,7 @@ public class Config {
 
     private static final ModConfigSpec.DoubleValue ALARM_VOLUME = BUILDER.comment("Volume of the alarm sound (0.0 - 1.0)").defineInRange("alarmVolume", 1.0, 0.0, 1.0);
 
-    private static final ModConfigSpec.IntValue SMOKE_DETECTOR_RADIUS = BUILDER.comment("Detection radius (in blocks) of the smoke detector. Fire within this radius will trigger the alarm and emit a redstone signal.").defineInRange("smokeDetectorRadius", 5, 1, 32);
+    private static final ModConfigSpec.IntValue SMOKE_DETECTOR_RADIUS = BUILDER.comment("Detection radius (in blocks) of the smoke detector. Fire within this radius will trigger the alarm and emit a redstone signal.").defineInRange("smokeDetectorRadius", 10, 1, 32);
 
     static final ModConfigSpec SPEC = BUILDER.build();